Changing input "q" to "resource_name"
Our members have used your ckanr package for a while and really enjoyed the awesome functions. One proposed change we have is regarding "q" in several search functions, since it is not intuitive and users can easily forget to include "name:". Perhaps we can add one line of code in the function so that users only need to input the name of resource. If this is preferable, I can help change the other functions with the same "q", as well as all the descriptions.

I'd also love to be more involved in helping improve the ckanr package by providing feedback and improvement suggestions. Please feel free to email me at mingbogo@gmail.com. Thank you.